pyproject.toml
[tool.poetry]
name = "receipt-reader"
version = "0.2.0"
description = "Read Sainsbury's receipts and upload them to Google Sheets"
authors = ["Alexandra Whitham"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^2.1.1"
configparser = "^7.1.0"
argparse = "^1.4.0"
pdf2image = "^1.17.0"
google-cloud-vision = "^3.7.4"
google-api-python-client = "^2.144.0"
google-auth-oauthlib = "^1.2.1"

[tool.poetry.dev-dependencies]
black = "^24.3.0"
isort = "^5.13"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
